requirements.txt
python-dotenv
fastapi
uvicorn[standard]
openai